+++ src/main/java/com/todo/Hiplanner/config/auth/PrincipalDetailsService.java
@@ -16,6 +16,7 @@ public class PrincipalDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        System.out.println("username = " + username);
         Member member = memberMapper.findByUsername(username);
         if(member != null){
             return new PrincipalDetails(member);
