+++ src/main/java/com/todo/Hiplanner/controller/MemberController.java
@@ -1,20 +1,17 @@
 package com.todo.Hiplanner.controller;
 
 import com.todo.Hiplanner.service.MemberService;
-import com.todo.Hiplanner.service.MemoService;
+import com.todo.Hiplanner.service.PlannerService;
 import com.todo.Hiplanner.vo.Member;
-import com.todo.Hiplanner.vo.Memo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.view.RedirectView;
 
 import javax.servlet.http.HttpSession;
-import java.time.LocalDate;
-import java.util.List;
 
 @Controller
 @RequestMapping("/members")
@@ -22,31 +19,8 @@ import java.util.List;
 public class MemberController {
 
     private final MemberService memberService;
-    private final MemoService memoService;
+    private final PlannerService memoService;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    @GetMapping("/logout")
-    public String logoutDo(HttpSession session){
-        session.removeAttribute("id");
-        return "member/login";
-    }
-
-    @PostMapping("/join")
-    public String joinProc(Member member){
-
-        member.setRole("ROLE_USER");
-        String passwordRaw = member.getPassword();
-        String encode = bCryptPasswordEncoder.encode(passwordRaw);
-        member.setPassword(encode);
-        memberService.insertMember(member);
-        return "redirect:/";
-    }
-
-    @ResponseBody
-    @PostMapping("/id/check")
-    public int idCheck(Member member){
-        return memberService.CheckId(member);
-    }
 
     @GetMapping("/{username}")
     public ModelAndView getInfo(ModelAndView mv, HttpSession session){
