+++ src/main/java/com/todo/Hiplanner/mapper/PlannerMapper.java
@@ -6,7 +6,7 @@ import org.apache.ibatis.annotations.Mapper;
 import java.util.List;
 
 @Mapper
-public interface plannerMapper {
+public interface PlannerMapper {
 
     public List<Memo> getMemo(Memo memo);
     public void insertMemo(Memo memo);
