--- src/main/java/com/todo/Hiplanner/repository/MemoRepository.java
@@ -1,7 +1,0 @@
-package com.todo.Hiplanner.repository;
-
-import com.todo.Hiplanner.vo.Memo;
-
-public interface MemoRepository {
-
-}
