--- src/main/java/com/todo/Hiplanner/repository/MemoRepositoryImpl.java
@@ -1,11 +1,0 @@
-package com.todo.Hiplanner.repository;
-
-import com.todo.Hiplanner.vo.Memo;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemoRepositoryImpl implements MemoRepository{
-
-
-}
