+++ src/main/java/com/todo/Hiplanner/service/PlannerServiceImpl.java
@@ -1,22 +1,19 @@
 package com.todo.Hiplanner.service;
 
-import com.todo.Hiplanner.mapper.plannerMapper;
+import com.todo.Hiplanner.mapper.PlannerMapper;
 import com.todo.Hiplanner.vo.Memo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.sql.Date;
-import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.Calendar;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class MemoServiceImpl implements MemoService{
+public class PlannerServiceImpl implements PlannerService {
 
-    private final plannerMapper plannerMapper;
+    private final PlannerMapper plannerMapper;
 
     @Override
     public List<Memo> getMemo(Memo memo) {
