+++ src/main/resources/mybatis-mapper/plannerMapper.xml
@@ -2,7 +2,7 @@
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.todo.Hiplanner.mapper.plannerMapper">
+<mapper namespace="com.todo.Hiplanner.mapper.PlannerMapper">
 
     <select id="getMemo" resultType="com.todo.Hiplanner.vo.Memo" parameterType="com.todo.Hiplanner.vo.Memo">
         SELECT TITLE, CONTENT, WDATE, BEGIN, END, PLANNO
